feat: episode begin and reward system
